gateway test refactor
